room-gen: move generator into scripts/ alongside import
It's an offline shard mutator like scrape-world (import), not runtime
engine machinery; engine/ otherwise holds only the tick subsystems.
